Continue if missing packages are in cache or available on kojipkgs (#6)
* Continue if missing packages are in cache
* Fallback missing packages from kojipkgs

+++ entrypoint.sh
@@ -2,6 +2,9 @@
 
 set -eu
 
+arch=${ARCH:-x86_64}
+
+
 if [ ! -f rpms.lock ]
 then
   echo "rpms.lock not found"
@@ -28,8 +31,12 @@ fi
 # Download rpms
 mkdir -p downloads
 cat rpms.lock | xargs -n256 dnf download \
-  --arch=x86_64 --arch=noarch \
-  --downloaddir=downloads
+  --skip-broken \
+  --arch=${arch} --arch=noarch \
+  --downloaddir=downloads \
+  2> >(tee download.err >&2)
+
+# Remove old packages
 while read package
 do
   if ! grep "^${package%.*.rpm}$" rpms.lock > /dev/null
@@ -38,6 +45,77 @@ do
   fi
 done < <(cd downloads; ls -1 *.rpm)
 
+# Check missing packages and download old packages from kojipkgs
+is_downloaded() {
+  if ls downloads/$1.* > /dev/null 2> /dev/null
+  then
+    return 0
+  fi
+  return 1
+}
+split_package_name() {
+  echo $1 | sed -n 's/\(\S\+\)-\([0-9a-zA-Z.-_]\+\)-\([0-9]\+\.fc[0-9]\+\)/\1 \2 \3/p'
+}
+warn_if_not_in_dnf_repo() {
+  if grep "^No package $1.* available.$" download.err > /dev/null 2> /dev/null
+  then
+    line_num=$(grep -n "$1" rpms.lock | cut -f1 -d:)
+    echo "::warning file=rpms.lock,line=${line_num},title=Package $1 missing from the repositories"
+  fi
+}
+
+line_num=-1
+error=false
+while read package
+do
+  (line_num+=1)
+
+  if is_downloaded ${package}
+  then
+    # Warn if not found in dnf repo but in local cache
+    warn_if_not_in_dnf_repo ${package}
+    continue
+  fi
+
+  echo "Downloading ${package} from kojipkgs"
+
+  pkg_fields=$(split_package_name ${package})
+  pkg_name=$(echo ${pkg_fields} | cut -f1 -d" ")
+  pkg_version=$(echo ${pkg_fields} | cut -f2 -d" ")
+  pkg_suffix=$(echo ${pkg_fields} | cut -f3 -d" ")
+
+  pkg_src=$(dnf info --available ${pkg_name} | sed -n 's/^Source\s*:\s*\(\S\+\).src.rpm/\1/p')
+  src_fields=$(split_package_name ${pkg_src})
+  src_name=$(echo ${src_fields} | cut -f1 -d" ")
+  for pkg_arch in ${arch} noarch
+  do
+    if wget -q \
+      https://kojipkgs.fedoraproject.org/packages/${src_name}/${pkg_version}/${pkg_suffix}/${pkg_arch}/${package}.${pkg_arch}.rpm \
+      -O downloads/${package}.${pkg_arch}.rpm
+    then
+      break
+    fi
+    rm downloads/${package}.${pkg_arch}.rpm
+  done
+
+  if is_downloaded ${package}
+  then
+    # Warn if not found in dnf repo and local cache but in kojipkgs
+    warn_if_not_in_dnf_repo ${package}
+    continue
+  fi
+
+  echo "::error file=rpms.lock,line=${line_num},title=Package ${package} unavailable"
+  error=true
+done < rpms.lock
+
+if ${error}
+then
+  echo "Missing packages" >&2
+  exit 1
+fi
+
+# Generate iso
 rm -rf iso-root/Packages
 mkdir -p iso-root/Packages
 while read rpm
